refactor: updates module refactor
